+++ Assets/Scripts/UI/GameUIManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 018988a82ab1f904588efd684e23bf48
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
