+++ Assets/Scripts/UI/TitleUIManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 014d998627b19104b9292fadb05353ec
+guid: 62f3b80887b7fba418aed003b720eae9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
